Add files via upload
Every circle has 8 octanes and the circle drawing algorithm generates all the points for one octane.
The points for other 7 octanes are generated by changing the sign towards X and Y coordinates.
MPC Algorithm plots the points along the circumference of the circle. But in order to fill the circle a line is drawn from the centre of the circle to the plotted point. This is done for every plotted point. Thus filling the circle while its being drawn.